Make event filtering a hook
