+++ src/utility/get-should-include-day.js
@@ -1,0 +1,10 @@
+function getShouldIncludeDay({ dayFilters, eventDate }) {
+  const dayOfWeekIndex = eventDate.getDay();
+  const { isActive } = dayFilters.find(function ({ ofWeek }) {
+    return ofWeek === dayOfWeekIndex;
+  });
+
+  return isActive;
+}
+
+export { getShouldIncludeDay };
